# api/openapi.yaml
openapi: 3.1.0
info:
  title: AgentOS API
  version: 0.1.0
  summary: Multi-tenant AI-agent infrastructure platform API
  description: |
    AgentOS is a multi-tenant AI-agent infrastructure platform. This contract
    covers the versioned REST API (/api/v1), the worker pull endpoint, the
    Server-Sent Events (SSE) execution-event stream, and platform health /
    metrics endpoints.

    ## Authentication
    Two schemes are accepted on every protected endpoint (see
    `securitySchemes`):
    1. `bearerToken` — `Authorization: Bearer <token>` where `<token>` is the
       HMAC-SHA256 JWT-style token returned by `POST /auth/login`
       (registration creates the first OWNER user; log in to get a token).
    2. `apiKey` — `X-API-Key: ak_...` header. API keys are created server-side
       (dev key created at startup) and authenticate as the key's OWNER role.
    Browser `EventSource` clients that cannot set headers may pass
    `?api_key=<key>` as a query parameter (legacy compatibility).

    ## RBAC
    Permissions: `agents.read`, `agents.write`, `runs.read`, `runs.execute`,
    `users.manage`, `organization.manage`. Roles: `OWNER` (all), `ADMIN`
    (all but organization.manage), `MEMBER` (agents.read, runs.read,
    runs.execute), `VIEWER` (agents.read, runs.read). Each operation declares
    its required permission via `x-required-permission`.

    ## Multi-tenancy
    Requests are scoped to an organization derived from the auth token. For
...

    ## Error model
...

...
servers:
...
security:
- bearerToken: []
- apiKey: []
paths:
...
  /runs:
    get:
      operationId: listRuns
      tags:
      - runs
      summary: List runs
      description: |
        Returns an envelope `{"runs":[...]}`. **Today this lists all runs
        across organizations** (handler TODO notes org filtering is future
        work). Unpaginated today.
      x-required-permission: runs.read
      x-unpaginated: true
      x-note: no organization filter applied yet
      responses:
        '200':
          description: Runs envelope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRunsResponse'
              example:
                runs:
                - ID: 3f6c2b62-2b57-4f8e-9e4e-2f4d3b6c9a10
                  OrganizationID: org-1
                  AgentID: agent-1
                  Input: What is 2+2?
                  Output: ''
                  Status: QUEUED
                  CreatedAt: '2025-01-01T12:00:00.123456789Z'
                  UpdatedAt: '2025-01-01T12:00:00.123456789Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: createRun
      tags:
      - runs
      summary: Create and enqueue a run
      description: |
        Creates a run (status `QUEUED`) and enqueues an `agent.run` task.
        Responds 201 with the run ID and `status: "queued"`. Actual execution
        happens in the worker; watch `GET /runs/{id}/events` (SSE) for
        `status.changed` events.
      x-required-permission: runs.execute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRunRequest'
            example:
              organization_id: org-1
              agent_id: agent-1
              input: What is 2+2?
      responses:
        '201':
          description: Run created and task queued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRunResponse'
              example:
                run_id: 3f6c2b62-2b57-4f8e-9e4e-2f4d3b6c9a10
                status: queued
        '400':
          description: Invalid body or missing agent_id.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/LegacyTextError'
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
...
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: HS256-HMAC
      description: |
        HMAC-SHA256 JWT-style token from `POST /auth/login` (24h expiry).
        Claims: user_id, organization_id, email, role, exp.
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Pre-provisioned API key (`ak_...`). Authenticates as the key's
        organization with the OWNER role. Browser EventSource clients may
        alternatively pass `?api_key=<key>` (legacy query-param auth).
  responses:
    Unauthorized:
      description: Missing/invalid credentials (plain text body today; JSON model is the target).
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/LegacyTextError'
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but lacking the required permission or organization access.
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/LegacyTextError'
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    LegacyTextError:
      type: string
      description: Legacy plain-text error body emitted via http.Error during migration.
      example: agent not found
    Error:
      type: object
      description: Structured error model (internal/httpx.ErrorModel).
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable stable error code (e.g. unauthorized, forbidden, not_found,
                bad_request, internal_error).
              examples:
              - not_found
            message:
              type: string
              description: Human-readable message.
              examples:
              - agent not found
            request_id:
              type: string
              description: Correlation ID echoed from the X-Request-ID header (or generated per request).
              examples:
              - 1b8f0c47e51d2a3c4d5e6f708192a3b4
          required:
          - code
          - message
          - request_id
      required:
      - error
...
    Run:
      type: object
      description: |
        Agent execution record (runs.Run; no json tags — PascalCase keys).
        Status transitions: QUEUED → RUNNING → COMPLETED | FAILED.
      properties:
        ID:
          type: string
          description: UUID v4.
          examples:
          - 3f6c2b62-2b57-4f8e-9e4e-2f4d3b6c9a10
        OrganizationID:
          type: string
          examples:
          - org-1
        AgentID:
          type: string
          examples:
          - agent-1
        Input:
          type: string
          examples:
          - What is 2+2?
        Output:
          type: string
          description: Final output once the run completes (empty until then).
          examples:
          - The result of 2+2 is 4
        Status:
          type: string
          enum:
          - QUEUED
          - RUNNING
          - COMPLETED
          - FAILED
        CreatedAt:
          type: string
          format: date-time
        UpdatedAt:
          type: string
          format: date-time
      required:
      - ID
      - OrganizationID
      - AgentID
      - Status
      - CreatedAt
      - UpdatedAt
...
    CreateRunRequest:
      type: object
      properties:
        organization_id:
          type: string
          description: Optional; defaults to the authenticated principal's organization.
          examples:
          - org-1
        agent_id:
          type: string
          examples:
          - agent-1
        input:
          type: string
          description: Prompt/task input handed to the agent.
          examples:
          - What is 2+2?
      required:
      - agent_id
      - input
    CreateRunResponse:
      type: object
      properties:
        run_id:
          type: string
          description: 'Created run ID (fallback: queue task ID if the runs service is unavailable).'
          examples:
          - 3f6c2b62-2b57-4f8e-9e4e-2f4d3b6c9a10
        status:
          type: string
          const: queued
      required:
      - run_id
      - status
    ListRunsResponse:
      type: object
      properties:
        runs:
          type: array
          items:
            $ref: '#/components/schemas/Run'
      required:
      - runs
...